+++ packages/compartment-mapper/src/import-archive.js
@@ -19,6 +19,8 @@ import { ZipReader } from '@endo/zip';
 import { link } from './link.js';
 import parserPreCjs from './parse-pre-cjs.js';
 import parserJson from './parse-json.js';
+import parserText from './parse-text.js';
+import parserBytes from './parse-bytes.js';
 import parserPreMjs from './parse-pre-mjs.js';
 import { parseLocatedJson } from './json.js';
 import { unpackReadPowers } from './powers.js';
@@ -38,6 +40,8 @@ const parserForLanguage = {
   'pre-cjs-json': parserPreCjs,
   'pre-mjs-json': parserPreMjs,
   json: parserJson,
+  text: parserText,
+  bytes: parserBytes,
 };
 
 /**